Create new files for EdgeHit implementations and pod update
